refactord persistence into package added sqlalchemy dependency added simple sqlalchemy class
