+++ requirements.txt
@@ -5,3 +5,4 @@ pyyaml
 google-cloud-vision
 pytesseract
 opencv-python
+sqlalchemy
